add unlocked burner logic
