+++ src/components/BurnerWalletCard/CreateFirstCastButton.tsx
@@ -9,7 +9,7 @@ export default function () {
   const firstCastMessage = `Create your ${used ? '' : 'first'} cast`
 
   return (
-    <Link href="/cast">
+    <Link href="/">
       <div className={displayTo('xs')}>
         <Button small type="primary">
           {firstCastMessage}
