+++ src/stores/BurnerInteractionStore.ts
@@ -1,0 +1,3 @@
+import { proxy } from 'valtio'
+
+export default proxy({ interactionClosed: true })
